requirements.txt
langchain-core>=0.3.30
langchain-community>=0.3.20
langchain-openai>=0.3.1
langgraph>=0.1.24
faiss-cpu>=1.10.0
python-dotenv>=1.0.1
pypdf>=5.0.0
sqlalchemy>=2.0.30